fix: use new format snapshotter config for containerd v2 (#91)
For some reason, despite still indicating a v2 config file,
the v3 syntax is necessary when specifying the snapshotter,
otherwise the default is used. Really this feels like a
containerd bug, but this approach appears to work.

Additionally, when using a non-default snapshotter, it
may be necessary to specify a platform, otherwise image
pulls can fail.

Signed-off-by: Andrew Bonney <andrew.bonney@bbc.co.uk>